+++ pywincffi/__init__.py
@@ -6,4 +6,4 @@ The root of the pywincffi package.  See the ``README`` and documentation for
 help and examples.
 """
 
-__version__ = (0, 5, 0)
+__version__ = (0, 6, 0)