+++ docs/source/changelog.rst
@@ -15,8 +15,8 @@ Notable enhancements and changes are:
 
     * :issue:`141` - :func:`pywincffi.kernel32.file.ReadFile` has been updated
       to provide improved support for overlapped IO. This is a **breaking change**
-      because the type of the returned value has changed and requires additional
-      processing.
+      because the return value and signature of the ReadFile function has
+      changed.
 
 0.5.0
 ~~~~~